refactor: move end-user CLI commands under infrahubctl
Instead of a separate `infrahub` entry point, register get, create,
update, and delete as top-level commands on the existing `infrahubctl`
CLI. Add schema list and schema show as subcommands of the existing
`infrahubctl schema` group alongside load, check, and export.

- Remove infrahub entry point from pyproject.toml
- Remove enduser_cli.py and enduser_commands.py
- Register commands in cli_commands.py
- Merge schema list/show into ctl/schema.py
- Update all tests to use cli_commands.app
