+++ infrahub_sdk/ctl/cli_commands.py
@@ -51,6 +51,10 @@ from ..template import Jinja2Template
 from ..template.exceptions import JinjaTemplateError
 from ..utils import write_to_file
 from ..yaml import SchemaFile
+from .commands.create import create_command
+from .commands.delete import delete_command
+from .commands.get import get_command
+from .commands.update import update_command
 from .exporter import dump
 from .importer import load
 from .parameters import CONFIG_PARAM
@@ -71,6 +75,10 @@ app.add_typer(task_app, name="task")
 
 app.command(name="dump")(dump)
 app.command(name="load")(load)
+app.command(name="get")(get_command)
+app.command(name="create")(create_command)
+app.command(name="update")(update_command)
+app.command(name="delete")(delete_command)
 
 console = Console()
