+++ infrahub_sdk/ctl/schema.py
@@ -10,12 +10,13 @@ import typer
 import yaml
 from pydantic import ValidationError
 from rich.console import Console
+from rich.table import Table
 
 from ..async_typer import AsyncTyper
 from ..ctl.client import initialize_client
 from ..ctl.utils import catch_exception, init_logging
 from ..queries import SCHEMA_HASH_SYNC_STATUS
-from ..schema import SchemaWarning
+from ..schema import NodeSchemaAPI, SchemaWarning
 from ..yaml import SchemaFile
 from .parameters import CONFIG_PARAM
 from .utils import load_yamlfile_from_disk_and_exit
@@ -258,3 +259,110 @@ async def export(
         console.print(f"[green] Exported namespace '{ns}' to {output_file}")
 
     console.print(f"[green] Schema exported to {directory}")
+
+
+@app.command(name="list")
+@catch_exception(console=console)
+async def schema_list(
+    filter_text: str | None = typer.Option(None, "--filter", help="Filter kinds by name substring"),
+    branch: str | None = typer.Option(None, "--branch", "-b", help="Target branch"),
+    _: str = CONFIG_PARAM,
+) -> None:
+    """List all available schema kinds.
+
+    Fetches the full schema from the Infrahub instance and displays a
+    table of node schema entries.  Use ``--filter`` to narrow results
+    by a case-insensitive substring match on the kind name.
+
+    Args:
+        filter_text: Optional substring to filter kind names.
+        branch: Target branch name.
+        _: Configuration file path (handled by callback).
+    """
+    client = initialize_client(branch=branch)
+    schemas = await client.schema.all(branch=branch)
+
+    items = list(schemas.values())
+    if filter_text:
+        items = [s for s in items if filter_text.lower() in s.kind.lower()]
+
+    items = [s for s in items if isinstance(s, NodeSchemaAPI)]
+    items.sort(key=lambda s: s.kind)
+
+    table = Table(title="Schema Kinds")
+    table.add_column("Namespace")
+    table.add_column("Name")
+    table.add_column("Kind")
+    table.add_column("Description")
+
+    for schema_item in items:
+        table.add_row(
+            schema_item.namespace,
+            schema_item.name,
+            schema_item.kind,
+            schema_item.description or "",
+        )
+
+    console.print(table)
+
+
+@app.command(name="show")
+@catch_exception(console=console)
+async def schema_show(
+    kind: str = typer.Argument(..., help="Schema kind to display"),
+    branch: str | None = typer.Option(None, "--branch", "-b", help="Target branch"),
+    _: str = CONFIG_PARAM,
+) -> None:
+    """Show details for a specific schema kind.
+
+    Displays metadata, attributes, and relationships for the requested
+    schema kind in a human-readable format.
+
+    Args:
+        kind: Infrahub schema kind (e.g. ``InfraDevice``).
+        branch: Target branch name.
+        _: Configuration file path (handled by callback).
+    """
+    client = initialize_client(branch=branch)
+    node_schema = await client.schema.get(kind=kind, branch=branch)
+
+    console.print(f"\n[bold]{node_schema.kind}[/bold]")
+    if node_schema.description:
+        console.print(f"  {node_schema.description}")
+    console.print(f"  Namespace: {node_schema.namespace}")
+    console.print(f"  Display Labels: {node_schema.display_labels or 'N/A'}")
+    console.print(f"  Human Friendly ID: {node_schema.human_friendly_id or 'N/A'}")
+
+    if node_schema.attributes:
+        attr_table = Table(title="Attributes")
+        attr_table.add_column("Name")
+        attr_table.add_column("Type")
+        attr_table.add_column("Required")
+        attr_table.add_column("Default")
+        attr_table.add_column("Description")
+
+        for attr in node_schema.attributes:
+            attr_table.add_row(
+                attr.name,
+                str(attr.kind),
+                "Yes" if not attr.optional else "No",
+                str(attr.default_value) if attr.default_value is not None else "",
+                attr.description or "",
+            )
+        console.print(attr_table)
+
+    if node_schema.relationships:
+        rel_table = Table(title="Relationships")
+        rel_table.add_column("Name")
+        rel_table.add_column("Peer")
+        rel_table.add_column("Cardinality")
+        rel_table.add_column("Optional")
+
+        for rel in node_schema.relationships:
+            rel_table.add_row(
+                rel.name,
+                rel.peer,
+                rel.cardinality,
+                "Yes" if rel.optional else "No",
+            )
+        console.print(rel_table)
